requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0
pydantic>=2.5.0
pyotp
qrcode
webauthn
httpx
geopy
user-agents
pillow
python-dotenv
